refactor: ActivityBar
