+++ packages/components/src/molecules/ActivityBar/ActivityBar.stories.tsx
@@ -1,13 +1,14 @@
-import { ComponentMeta, ComponentStory } from "@storybook/react";
-import { ActivityBar } from "./ActivityBar";
-import { ExtraActivityBarArgs, HiddenOptionsActivityBarArgs, MainActivityBarArgs } from "./args";
+import {ComponentMeta, ComponentStory} from '@storybook/react';
+import ActivityBar from './ActivityBar';
+
+import {ExtraActivityBarArgs, HiddenOptionsActivityBarArgs, MainActivityBarArgs} from './args';
 
 export default {
-  title: "Molecules/ActivityBar",
+  title: 'Molecules/ActivityBar',
   component: ActivityBar,
 } as ComponentMeta<typeof ActivityBar>;
 
-const Template: ComponentStory<typeof ActivityBar> = (args) => <ActivityBar {...args} />;
+const Template: ComponentStory<typeof ActivityBar> = args => <ActivityBar {...args} />;
 
 export const MainActivityBar = Template.bind({});
 MainActivityBar.args = MainActivityBarArgs;
