+++ packages/components/src/molecules/WalkThroughCard/index.ts
@@ -1,1 +1,2 @@
 export {default as WalkThroughCard} from './WalkThroughCard';
+export type {WalkThroughCardProps} from './types';
